requirements.txt
python-telegram-bot>=20.0
python-dotenv
pymongo>=4.10.1
dnspython>=2.7.0